requirements.txt
playwright
streamlit
Pillow
requests
pytesseract
pyobjc-framework-Vision; sys_platform == 'darwin'